Move StateKeyTuplerSorter to the types package
